+++ examples/data-sources/azureipam_block_networks_availables/data-source.tf
@@ -1,4 +1,4 @@
-# Returns one block with network and usage information
+# Returns the list of resource Ids of available networks to be associated to a space block
 data "azureipam_block_networks_availables" "example" {
   space = "au"
   block = "AustraliaEast"