+++ tests/block_networks_data_source/main.tf
@@ -1,0 +1,8 @@
+# Returns the list of networks actively associated to a space block
+data "azureipam_block_networks" "example" {
+  space = "au"
+  block = "AustraliaEast"
+}
+output "associated" {
+  value = data.azureipam_block_networks.example
+}